+++ src/main/kotlin/net/pwall/json/kotlin/codegen/gradle/extension/SchemaExtension.kt
@@ -1,5 +1,5 @@
 /*
- * @(#) ClassMapping.kt
+ * @(#) SchemaExtension.kt
  *
  * json-kotlin-gradle  Gradle Code Generation Plugin for JSON Schema
  * Copyright (c) 2021 Peter Wall
@@ -23,8 +23,9 @@
  * SOFTWARE.
  */
 
-package net.pwall.json.kotlin.codegen.gradle.mapping
+package net.pwall.json.kotlin.codegen.gradle.extension
 
+import java.net.URI
 import javax.inject.Inject
 
 import org.gradle.api.Named
@@ -32,27 +33,19 @@ import org.gradle.api.Project
 import org.gradle.api.tasks.Input
 import org.gradle.kotlin.dsl.property
 
-import net.pwall.json.schema.codegen.ClassId
-import net.pwall.json.schema.codegen.ClassName
-import net.pwall.json.schema.codegen.CodeGenerator
+import net.pwall.json.pointer.JSONPointer
+import net.pwall.json.schema.JSONSchema
 
-@Suppress("UnstableApiUsage")
-abstract class ClassMapping @Inject constructor(@Input val name0: String, project: Project) : Named {
+abstract class SchemaExtension @Inject constructor(@Input val name0: String, project: Project) : Named {
 
     override fun getName(): String = name0
 
     @Input
-    val className = project.objects.property<String>()
+    val keyword = project.objects.property<String>()
 
-    abstract fun applyTo(codeGenerator: CodeGenerator)
+    @Input
+    val value = project.objects.property<String>()
 
-    val classId: ClassId
-        get() = className.get().let {
-            val i = it.lastIndexOf('.')
-            if (i < 0)
-                ClassName(it)
-            else
-                ClassName(it.substring(i + 1), it.substring(0, i))
-        }
+    abstract fun validator(uri: URI?, pointer: JSONPointer): JSONSchema.Validator
 
 }